backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.4.2
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.38.0
openai==1.1.1
python-dotenv==1.0.0
httpx==0.25.0
pytest==7.4.2
pytest-asyncio==0.21.1
python-multipart==0.0.6
sqlalchemy==2.0.21
psycopg2-binary==2.9.9
alembic==1.12.0
